+++ src/lib/components/page/PaginatorPage.tsx
@@ -11,88 +11,91 @@ import { RENDERING_STATES } from '../../../utils/enums'
  */
 
 const PaginatorPage = () => {
-  const { pdf, activePage, scale, rerender, setRendering, setDesiredScale } =
-    useDocumentContext()
+  const {
+    pdf,
+    activePage,
+    scale,
+    rerender,
+    setPaginatorPageRender,
+    setDesiredScale,
+  } = useDocumentContext()
 
   /**
    * Renders the page and all its layers
    *
    * @returns A promise that resolves when the page is rendered
    */
-  const renderPage = useCallback(async () => {
-    setRendering(RENDERING_STATES.RENDERING)
+  const renderPage = useCallback(
+    async (viewer: HTMLElement) => {
+      return await new Promise((resolve) => {
+        pdf?.getPage(activePage).then(async (page) => {
+          const height = document
+            .getElementById('evilFlowersContent')!
+            .getBoundingClientRect().height
 
-    return await new Promise((resolve) => {
-      pdf?.getPage(activePage).then((page) => {
-        const container = document.createElement('textLayer')
-        container.setAttribute('id', 'textLayer')
-        container.setAttribute('class', 'pdf-canvas-textLayer')
+          let viewport = page.getViewport({ scale })
+          const calcScreenHeight = 0.95 * height
+          const desiredHeight = calcScreenHeight * viewport.scale
+          const viewportHeight = viewport.height / viewport.scale
+          const desiredScale = desiredHeight / viewportHeight
+          setDesiredScale(desiredScale)
+          viewport = page.getViewport({ scale: desiredScale })
 
-        let viewport = page.getViewport({ scale })
-        const calcScreenHeight = 0.95 * (window.innerHeight - 210)
-        const desiredHeight = calcScreenHeight * viewport.scale
-        const viewportHeight = viewport.height / viewport.scale
-        const desiredScale = desiredHeight / viewportHeight
-        setDesiredScale(desiredScale)
-        viewport = page.getViewport({ scale: desiredScale })
+          const container = document.createElement('textLayer')
+          container.setAttribute('id', 'textLayer')
+          container.setAttribute('class', 'pdf-canvas-textLayer')
 
-        page.getTextContent().then((textContent) => {
-          pdfjs.renderTextLayer({
-            textContent,
-            container,
-            viewport,
-            textDivs: [],
+          page.getTextContent().then((textContent) => {
+            pdfjs.renderTextLayer({
+              textContent,
+              container,
+              viewport,
+              textDivs: [],
+            })
+          })
+
+          const canvas = document.createElement('canvas')
+          canvas.setAttribute('id', 'evilFlowersCanvas')
+          canvas.setAttribute('class', 'pdf-canvas-container-paginator')
+          canvas.width = viewport.width
+          canvas.height = viewport.height
+
+          const context = canvas.getContext('2d')
+
+          const renderContext = {
+            canvasContext: context as Object,
+            viewport: viewport,
+          }
+          const renderTask = page.render(renderContext)
+          await renderTask.promise.then(() => {
+            viewer.replaceChildren(canvas, container)
+
+            resolve('')
           })
         })
-
-        const canvas =
-          (document.getElementById('evilFlowersCanvas') as HTMLCanvasElement) ??
-          document.createElement('canvas')
-        canvas.setAttribute('id', 'evilFlowersCanvas')
-        canvas.width = viewport.width
-        canvas.height = viewport.height
-        canvas.style.width = viewport.width + 'px'
-        canvas.style.height = viewport.height + 'px'
-        const context = canvas.getContext('2d')
-
-        const renderContext = {
-          canvasContext: context as Object,
-          viewport: viewport,
-        }
-        const renderTask = page.render(renderContext)
-        renderTask.promise.then(() => {
-          const prevCanvas = document.getElementById(
-            'evilFlowersCanvas'
-          ) as HTMLCanvasElement
-          const prevTextLayerNode = document.getElementById(
-            'textLayer'
-          ) as HTMLElement
-
-          if (prevTextLayerNode)
-            document
-              .getElementById('evilFlowersContent')
-              ?.replaceChild(container, prevTextLayerNode)
-          else
-            document
-              .getElementById('evilFlowersContent')
-              ?.replaceChildren(container)
-          if (!prevCanvas)
-            document.getElementById('evilFlowersContent')?.appendChild(canvas)
-
-          resolve(RENDERING_STATES.RENDERED)
-        })
       })
-    })
-  }, [activePage, pdf, scale])
+    },
+    [activePage, pdf, scale]
+  )
 
   useEffect(() => {
-    console.log('ahoj')
-    renderPage().then((resolve) => {
-      setRendering(RENDERING_STATES.RENDERED)
+    setPaginatorPageRender(RENDERING_STATES.RENDERING)
+    const loader = document.createElement('div')
+    loader.setAttribute('class', 'viewer-loader-small')
+
+    const loadPage = async () => {
+      const viewer = document.getElementById('evilFlowersPaginatorContent')!
+      viewer.replaceChildren()
+      viewer.appendChild(loader)
+
+      await renderPage(viewer)
+    }
+    loadPage().then(() => {
+      setPaginatorPageRender(RENDERING_STATES.RENDERED)
     })
   }, [activePage, pdf, scale, rerender])
 
-  return <div id={'evilFlowersContent'} className={'pdf-container'}></div>
+  return <div id={'evilFlowersPaginatorContent'}></div>
 }
 
 export default PaginatorPage