+++ src/lib/components/share/ShareQRCode.tsx
@@ -73,7 +73,12 @@ const ShareQRCode = (props: IShareQRProps) => {
       <div className={'share-qr-code-container'}>
         {props.link ? (
           props.link !== 'error' ? (
-            <QRCode id="QRCode" size={250} value={props.link} />
+            <QRCode
+              style={{ border: '5px solid white' }}
+              id="QRCode"
+              size={250}
+              value={props.link}
+            />
           ) : (
             <pre className="share-failed-qr-code-text">
               {t('shareQrCodeFail')}